+++ src/Const.js
@@ -35,6 +35,8 @@ export default {
     DATE: 'DateFilter',
     CUSTOM: 'CustomFilter'
   },
+  FILTER_COND_EQ: 'eq',
+  FILTER_COND_LIKE: 'like',
   EXPAND_BY_ROW: 'row',
   EXPAND_BY_COL: 'column'
 };