+++ src/TableRow.js
@@ -39,22 +39,23 @@ class TableRow extends Component {
     if (e.target.tagName !== 'INPUT' &&
         e.target.tagName !== 'SELECT' &&
         e.target.tagName !== 'TEXTAREA') {
-      const rowIndex = e.currentTarget.rowIndex + 1;
       if (this.props.onRowDoubleClick) {
-        this.props.onRowDoubleClick(rowIndex);
+        this.props.onRowDoubleClick(this.props.index);
       }
     }
   }
 
   rowMouseOut = e => {
+    const rowIndex = this.props.index;
     if (this.props.onRowMouseOut) {
-      this.props.onRowMouseOut(e.currentTarget.rowIndex, e);
+      this.props.onRowMouseOut(rowIndex, e);
     }
   }
 
   rowMouseOver = e => {
+    const rowIndex = this.props.index;
     if (this.props.onRowMouseOver) {
-      this.props.onRowMouseOver(e.currentTarget.rowIndex, e);
+      this.props.onRowMouseOver(rowIndex, e);
     }
   }